'fix(nvim): isolate sidekick window lifecycle'
